src/auth.py
"""
User authentication and session management
"""
import streamlit_authenticator as stauth
import yaml
from yaml.loader import SafeLoader

class AuthManager:
    """Handles user authentication"""
    
    def __init__(self, config_path='config.yaml'):
        self.config_path = config_path
        self.authenticator = None
        
    def load_config(self):
        """Load authentication configuration"""
        try:
            with open(self.config_path) as file:
                config = yaml.load(file, Loader=SafeLoader)
            return config
        except FileNotFoundError:
            return None
    
    def initialize_authenticator(self):
        """Initialize the authenticator"""
        # Implementation will be added in later steps
        pass
